# Conflicts:
#   frontend/bgs/src/pages/info/FollowerFollowingListPage.jsx
#   frontend/bgs/src/pages/info/MyInfoEditPage.jsx
#   frontend/bgs/src/pages/info/MyInfoPage.jsx
#   frontend/bgs/src/pages/info/MyInfoViewPage.jsx

+++ backend/bgs/src/main/java/com/ssafy/bgs/auth/service/SocialUserResponseDto.java
@@ -18,8 +18,8 @@ public class SocialUserResponseDto {
     private String nickname;
     private LocalDate birthDate;
     private String sex;
-    private Integer height;
-    private Integer weight;
+    private Double height;
+    private Double weight;
     private Double degree;
     private String introduction;
     private Double totalWeight;
